Enormous MUST DELETE. Removed cond_waits and cond_signals from ServiceThread ServiceThread now uses semaphores for sleeping and waking up. Access to fSubmittedList is protected with fSubmittedMutex. ServiceThread now locks on fServiceMutex. Also used by the TMLock.

+++ Sources/ooptaskdependlist.cpp
@@ -87,7 +87,7 @@ OOPDataVersion & OOPTaskDependList::Version(int i){
   return fDependList[i].fVersion;
 }
 TPZAutoPointer<TPZSaveable> OOPTaskDependList::ObjPtr(int i){
-  if(i< fDependList.size())
+  if((unsigned)i< fDependList.size())
   {
 #ifdef LOGPZ
